Optimization: only create a single FileFinder for the data directory.
Create an instance of the FileFinder in OMW::Engine and pass as necessary
instead of creating an instance when needed, since construction is expensive.
The disadvantage of this is that a reasonably large table is kept in memory
despite not being needed after initialization.

(In my case went from 8e9 to 3e8 operations to construct FileFinder instances.
Note: my data dir has 11606 files, so this is an extreme example.)